+++ lib/evil_events/core/events.rb
@@ -8,6 +8,8 @@ module EvilEvents::Core
     require_relative 'events/serializers/base'
     require_relative 'events/serializers/hash'
     require_relative 'events/serializers/json'
+    require_relative 'events/serializers/xml'
+    require_relative 'events/serializers/xml/event_serialization_state'
     require_relative 'events/event_extensions/type_aliasing'
     require_relative 'events/event_extensions/payloadable'
     require_relative 'events/event_extensions/payloadable/abstract_payload'