+++ lib/evil_events/serializer.rb
@@ -18,6 +18,13 @@ module EvilEvents
       def load_from_hash(serialized_event)
         EvilEvents::Core::Bootstrap[:event_system].deserialize_from_hash(serialized_event)
       end
+
+      # @see EvilEvents::Core::System
+      # @api public
+      # @since 0.4.0
+      def load_from_xml(serialized_event)
+        EvilEvents::Core::Bootstrap[:event_system].deserialize_from_xml(serialized_event)
+      end
     end
   end
 end